provide fallback support for non fine-grained rbac vm actions

+++ frontend/src/routes/Infrastructure/VirtualMachines/utils.tsx
@@ -8,6 +8,11 @@ import { ActionExtensionProps, ListColumnExtensionProps } from '../../../plugin-
 import { IResourceDefinition } from '../../../resources'
 import { Cluster } from '../../../resources/utils'
 import { IAcmRowAction, IAcmTableColumn } from '../../../ui-components'
+import {
+  ClosedDeleteExternalResourceModalProps,
+  IDeleteExternalResourceModalProps,
+} from '../../Search/components/Modals/DeleteExternalResourceModal'
+import { ClosedDeleteModalProps, IDeleteModalProps } from '../../Search/components/Modals/DeleteResourceModal'
 import { GetUrlSearchParam } from '../../Search/searchDefinitions'
 import { ClosedVMActionModalProps, IVMActionModalProps } from './modals/VMActionModal'
 
@@ -76,7 +81,10 @@ export const printableVMStatus = {
 export function getVirtualMachineRowActions(
   item: any,
   allClusters: Cluster[],
+  setDeleteResource: Dispatch<SetStateAction<IDeleteModalProps>>,
+  setDeleteExternalResource: Dispatch<SetStateAction<IDeleteExternalResourceModalProps>>,
   setVMAction: Dispatch<SetStateAction<IVMActionModalProps>>,
+  isFineGrainedRbacEnabled: boolean,
   vmActionsEnabled: boolean,
   navigate: NavigateFunction,
   t: TFunction<string, undefined>,
@@ -138,13 +146,30 @@ export function getVirtualMachineRowActions(
     id: 'delete',
     title: t('Delete VirtualMachine'),
     click: (item: any) => {
-      setVMAction({
-        open: true,
-        close: () => setVMAction(ClosedVMActionModalProps),
-        action: 'Delete',
-        method: 'DELETE',
-        item,
-      })
+      if (isFineGrainedRbacEnabled) {
+        setVMAction({
+          open: true,
+          close: () => setVMAction(ClosedVMActionModalProps),
+          action: 'Delete',
+          method: 'DELETE',
+          item,
+        })
+      } else if (item.managedHub && item.managedHub !== 'global-hub') {
+        setDeleteExternalResource({
+          open: true,
+          close: () => setDeleteExternalResource(ClosedDeleteExternalResourceModalProps),
+          resource: item,
+          hubCluster: allClusters.find((cluster) => cluster.name === item.managedHub),
+        })
+      } else {
+        setDeleteResource({
+          open: true,
+          close: () => setDeleteResource(ClosedDeleteModalProps),
+          resource: item,
+          currentQuery: 'kind:VirtualMachine,VirtualMachineInstance',
+          relatedResource: false,
+        })
+      }
     },
     isDisabled: isVMDeleteDisabled(printableStatus, item),
     tooltip: isVMDeleteDisabled(printableStatus, item)
@@ -261,6 +286,9 @@ export function getVMSnapshotActions(
   isVMRunning: boolean,
   allClusters: Cluster[],
   vmActionsEnabled: boolean,
+  isFineGrainedRbacEnabled: boolean,
+  setDeleteResource: Dispatch<SetStateAction<IDeleteModalProps>>,
+  setDeleteExternalResource: Dispatch<SetStateAction<IDeleteExternalResourceModalProps>>,
   setVMAction: Dispatch<SetStateAction<IVMActionModalProps>>,
   navigate: NavigateFunction,
   t: TFunction<string, undefined>
@@ -330,13 +358,30 @@ export function getVMSnapshotActions(
     id: 'delete',
     title: t('Delete VirtualMachineSnapshot'),
     click: (item: any) => {
-      setVMAction({
-        open: true,
-        close: () => setVMAction(ClosedVMActionModalProps),
-        action: 'Delete',
-        method: 'DELETE',
-        item,
-      })
+      if (isFineGrainedRbacEnabled) {
+        setVMAction({
+          open: true,
+          close: () => setVMAction(ClosedVMActionModalProps),
+          action: 'Delete',
+          method: 'DELETE',
+          item,
+        })
+      } else if (item.managedHub && item.managedHub !== 'global-hub') {
+        setDeleteExternalResource({
+          open: true,
+          close: () => setDeleteExternalResource(ClosedDeleteExternalResourceModalProps),
+          resource: item,
+          hubCluster: allClusters.find((cluster) => cluster.name === item.managedHub),
+        })
+      } else {
+        setDeleteResource({
+          open: true,
+          close: () => setDeleteResource(ClosedDeleteModalProps),
+          resource: item,
+          currentQuery: 'kind:VirtualMachineSnapshot',
+          relatedResource: false,
+        })
+      }
     },
   }
